// 3-frontend/src/app/core/models/parametro/request/actualizar-negocio-request.model.ts
export class ActualizarNegocioRequest {
    idUsuario: string;
    idMoneda: number;
    codMoneda: string;
    destinationTimeZoneIdActualizacion: string;
    razonSocial: string;
    ruc: string;
    direccion: string;
    correoElectronico: string;
    celular: string;
    correo: string;
    nombreArchivo: string;
    foto: string;
    colorBoleta: string;
    codFormatoImpresion: string;
}